requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
--extra-index-url https://test.pypi.org/simple/
monpetitsalon==0.7.0
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
lxml==5.1.0 ; python_version >= "3.11" and python_version < "4.0"
outcome==1.3.0.post0 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.11" and python_version < "4.0"
pendulum==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and os_name == "nt" and implementation_name != "pypy" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pytzdata==2020.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
selenium==4.18.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.11" and python_version < "4.0"
trio-websocket==0.11.1 ; python_version >= "3.11" and python_version < "4.0"
trio==0.25.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3[socks]==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
webdriver-manager==4.0.1 ; python_version >= "3.11" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pymongo==4.6.3 ; python_version >= "3.11" and python_version < "4.0"
apache-airflow-providers-mongo==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
